requirements.txt
fastapi
uvicorn
python-docx
pymupdf
requests
python-dotenv
chromadb
python-multipart
